conclui as implementacoes de contadores

+++ slSiva/presenter/Contador.cs
@@ -3,7 +3,7 @@
     public class Contador
     {
         public decimal SQ_CONTADOR { get; set; }
-        public long NU_CRC { get; set; }
+        public decimal NU_CRC { get; set; }
         public string NM_CONTADOR { get; set; }
         public string NM_LOGRADOURO { get; set; }
         public string NU_LOGRADOURO { get; set; }
@@ -12,6 +12,5 @@
         public string NM_BAIRRO { get; set; }
         public string NU_TELEFONE { get; set; }
         public string NM_EMAIL_COMERCIAL { get; set; }
-
     }
 }
